backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
psycopg2-binary>=2.9.9
httpx>=0.27.0
boto3>=1.34.0
PyYAML>=6.0.1
django-unfold>=0.20.0
drf-spectacular>=0.27.0
djangorestframework-simplejwt>=5.3.1